Huge Changes! Fixed Main Bug Of User Not Getting Re-Authenticated When They Update Their Profile DP and Later Try To Login + Re-Structured All The UI Via Tailwind CSS + Daisy UI.. Bootstrap Styling Has Been Removed Completely + Fixed Bug Of Adding New Redundant User Role To Database Whenever User Get's Registered + Re-Structured Admin Panel - Changes In The UI and Fixed Bug Of Uploading Pictures Where They Get Uploaded In Target's Static Folder and Gets Deleted After Project Restart.. Now, The Images Uploaded From Website Will Be Store Permanently In The Original Static Folder + Some Minor Bug Fixes

+++ src/main/java/com/ganga/food_app/controller/AddressController.java
@@ -47,7 +47,6 @@ public class AddressController {
     @PostMapping("/create")
     public String createAddress(@Valid @ModelAttribute("address") Address address, BindingResult result, Model model, HttpSession session) {
         if (result.hasErrors()) {
-            System.out.println(result.getFieldErrors());
             return "addr/addrForm";
         }
 
@@ -64,8 +63,12 @@ public class AddressController {
 
     @PostMapping("/delete")
     public String deleteAddress(@RequestParam("addressId") UUID id, HttpSession session) {
-        addressService.removeAddress(id);
-        session.setAttribute("message", new Message("Address Deleted Successfully", MessageType.SUCCESS));
+        if(addressService.isAddressInUse(id) != null) {
+            session.setAttribute("message", new Message("This address is in use with atleast one of your Orders.. It Cannot Be Deleted..", MessageType.DANGER));
+        } else {
+            addressService.removeAddress(id);
+            session.setAttribute("message", new Message("Address Deleted Successfully", MessageType.SUCCESS));
+        }
         return "redirect:/address";
     }
 }
